Make 'make_app' work with generic product

+++ make_app.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 #
-# Wrap product-snsproduct-sns/target/product-sns-4.6.0-mac.zip
+# Wrap product-sns/target/product-sns-4.6.0-mac.zip
+# or ../phoebus/phoebus-product/target/phoebus-4.6.3-mac.zip
 # as Mac OS X CSS_Phoebus.app with JDK based on $JAVA_HOME
 #
 # phoebus.app/jdk
@@ -14,7 +15,25 @@
 #
 # Author: Kay Kasemir
 
+if [ $# -eq 1 ]
+then
+  PROD=$1
+else
+  PROD=`echo product-sns/target/product-sns-*-mac.zip`
+fi
+
+# Leaving original ZIP, creating new ZIP here
+DEST=`basename $PROD`
+echo "Turning $PROD into Mac OS app $DEST"
+
+if [ "$PROD" = "$DEST" ]
+then
+  echo "$PROD must be in different directory than $DEST"
+  exit 1
+fi
+
 APP=CSS_Phoebus.app
+
 # *.app skeleton w/ launch script
 rm -rf $APP
 cp -r app_template $APP
@@ -27,23 +46,20 @@ then
   cp -r $JDK $APP/jdk
 else
   echo "Missing $JAVA_HOME set to JDK/Contents/Home"
-  exit 1
+  exit 2
 fi
 
 # Add product
-PROD=`echo product-sns/target/product-sns-*-mac.zip`
 if [ -r "$PROD" ]
 then
   echo "Adding $PROD"
   unzip -q $PROD -d $APP
 else
   echo "Cannot locate product-sns-*-mac.zip"
-  exit 2
+  exit 3
 fi
 
-V=`echo $PROD | egrep -o '[0-9.]+' | head -n1`
-
-echo "Packing product-sns-${V}-mac.zip"
-rm -f product-sns-${V}-mac.zip
-zip -qr product-sns-${V}-mac.zip $APP
+echo "Packing $APP as $DEST"
+rm -f $DEST
+zip -qr $DEST $APP
 # rm -rf $APP
